+++ backend/src/main/java/com/bakdata/conquery/models/config/SqlConnectorConfig.java
@@ -1,0 +1,9 @@
+package com.bakdata.conquery.models.config;
+
+import lombok.Data;
+
+@Data
+public class SqlConnectorConfig {
+
+	boolean enabled;
+}
