+++ backend/src/main/java/com/bakdata/conquery/models/worker/ShardConnectionManager.java
@@ -1,0 +1,112 @@
+package com.bakdata.conquery.models.worker;
+
+import java.net.InetSocketAddress;
+import java.util.function.Function;
+
+import javax.validation.Validator;
+
+import com.bakdata.conquery.io.jackson.View;
+import com.bakdata.conquery.io.mina.BinaryJacksonCoder;
+import com.bakdata.conquery.io.mina.CQProtocolCodecFilter;
+import com.bakdata.conquery.io.mina.ChunkReader;
+import com.bakdata.conquery.io.mina.ChunkWriter;
+import com.bakdata.conquery.io.mina.MinaAttributes;
+import com.bakdata.conquery.io.mina.NetworkSession;
+import com.bakdata.conquery.models.config.ConqueryConfig;
+import com.bakdata.conquery.models.jobs.Job;
+import com.bakdata.conquery.models.jobs.JobManager;
+import com.bakdata.conquery.models.jobs.ReactingJob;
+import com.bakdata.conquery.models.messages.SlowMessage;
+import com.bakdata.conquery.models.messages.network.MessageToManagerNode;
+import com.bakdata.conquery.models.messages.network.NetworkMessageContext;
+import com.bakdata.conquery.util.io.ConqueryMDC;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.mina.core.service.IoAcceptor;
+import org.apache.mina.core.service.IoHandlerAdapter;
+import org.apache.mina.core.session.IoSession;
+import org.apache.mina.transport.socket.nio.NioSocketAcceptor;
+
+/**
+ * Manager of the connection from the manager to the ConQuery shards.
+ */
+@Slf4j
+@RequiredArgsConstructor
+public class ShardConnectionManager extends IoHandlerAdapter implements ClusterManager {
+
+	private IoAcceptor acceptor;
+	private final DistributedDatasetRegistry datasetRegistry;
+	private final JobManager jobManager;
+	private final Validator validator;
+	private final ConqueryConfig config;
+	private final Function<Class<? extends View>, ObjectMapper> createInternalObjectMapperCreator;
+
+	@Override
+	public void sessionOpened(IoSession session) {
+		ConqueryMDC.setLocation("ManagerNode[" + session.getLocalAddress().toString() + "]");
+		log.info("New client {} connected, waiting for identity", session.getRemoteAddress());
+	}
+
+	@Override
+	public void sessionClosed(IoSession session) {
+		ConqueryMDC.setLocation("ManagerNode[" + session.getLocalAddress().toString() + "]");
+		log.info("Client '{}' disconnected ", session.getAttribute(MinaAttributes.IDENTIFIER));
+	}
+
+	@Override
+	public void exceptionCaught(IoSession session, Throwable cause) {
+		ConqueryMDC.setLocation("ManagerNode[" + session.getLocalAddress().toString() + "]");
+		log.error("caught exception", cause);
+	}
+
+	@Override
+	public void messageReceived(IoSession session, Object message) {
+		ConqueryMDC.setLocation("ManagerNode[" + session.getLocalAddress().toString() + "]");
+		if (message instanceof MessageToManagerNode toManagerNode) {
+
+			log.trace("ManagerNode received {} from {}", message.getClass().getSimpleName(), session.getRemoteAddress());
+
+			Job job = new ReactingJob<>(toManagerNode,
+				new NetworkMessageContext.ManagerNodeNetworkContext(new NetworkSession(session),
+					datasetRegistry,
+					config.getCluster().getBackpressure()));
+
+			if (toManagerNode instanceof SlowMessage slowMessage) {
+				slowMessage.setProgressReporter(job.getProgressReporter());
+				jobManager.addSlowJob(job);
+			}
+			else {
+				jobManager.addFastJob(job);
+			}
+		}
+		else {
+			log.error("Unknown message type {} in {}", message.getClass(), message);
+		}
+	}
+
+	@Override
+	public void start() throws Exception {
+		acceptor = new NioSocketAcceptor();
+
+		ObjectMapper om = createInternalObjectMapperCreator.apply(View.InternalCommunication.class);
+		config.configureObjectMapper(om);
+		BinaryJacksonCoder coder = new BinaryJacksonCoder(datasetRegistry, validator, om);
+		acceptor.getFilterChain().addLast("codec", new CQProtocolCodecFilter(new ChunkWriter(coder), new ChunkReader(coder, om)));
+		acceptor.setHandler(this);
+		acceptor.getSessionConfig().setAll(config.getCluster().getMina());
+		acceptor.bind(new InetSocketAddress(config.getCluster().getPort()));
+		log.info("Started ManagerNode @ {}", acceptor.getLocalAddress());
+	}
+
+	@Override
+	public void stop() throws Exception {
+		try {
+			acceptor.dispose();
+		}
+		catch (Exception e) {
+			log.error(acceptor + " could not be closed", e);
+		}
+
+	}
+}
