+++ backend/src/main/java/com/bakdata/conquery/resources/admin/rest/AdminConceptsResource.java
@@ -29,7 +29,7 @@ import lombok.Setter;
 @RequiredArgsConstructor(onConstructor_ = {@Inject})
 public class AdminConceptsResource {
 
-	private final AdminDatasetProcessor processor;
+	private final AdminDatasetProcessor<Namespace> processor;
 
 	@PathParam(DATASET)
 	private Dataset dataset;
