+++ backend/src/main/java/com/bakdata/conquery/resources/admin/rest/AdminDatasetResource.java
@@ -57,7 +57,7 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor(onConstructor_ = {@Inject})
 public class AdminDatasetResource {
 
-	private final AdminDatasetProcessor processor;
+	private final AdminDatasetProcessor<Namespace> processor;
 
 	@PathParam(DATASET)
 	private Dataset dataset;
