+++ backend/src/main/java/com/bakdata/conquery/resources/admin/ui/model/UIContext.java
@@ -2,7 +2,7 @@ package com.bakdata.conquery.resources.admin.ui.model;
 
 import java.util.Collection;
 
-import com.bakdata.conquery.models.worker.DatasetRegistry;
+import com.bakdata.conquery.models.worker.DistributedDatasetRegistry;
 import com.bakdata.conquery.models.worker.WorkerInformation;
 import com.bakdata.conquery.resources.ResourceConstants;
 import freemarker.template.TemplateModel;
@@ -15,7 +15,7 @@ public class UIContext {
 	private static final TemplateModel STATIC_URI_ELEMENTS = ResourceConstants.getAsTemplateModel();
 
 	@Getter
-	private final DatasetRegistry namespaces;
+	private final DistributedDatasetRegistry namespaces;
 
 	@Getter
 	public final TemplateModel staticUriElem = STATIC_URI_ELEMENTS;
@@ -23,13 +23,15 @@ public class UIContext {
 	public boolean[] getWorkerStatuses() {
 		boolean[] result = new boolean[namespaces.getShardNodes().values().size()];
 		int id = 0;
-		for(WorkerInformation wi:namespaces.getWorkers().values()) {
+		for (WorkerInformation wi : this.getWorkers()) {
 			result[id++] = wi.isConnected();
 		}
 		return result;
 	}
 	
 	public Collection<WorkerInformation> getWorkers() {
-		return namespaces.getWorkers().values();
+		return namespaces.getDatasets().stream()
+			.flatMap(ns -> ns.getWorkerHandler().getWorkers().values().stream())
+			.toList();
 	}
 }
