+++ backend/src/main/java/com/bakdata/conquery/sql/conquery/SqlExecutionManager.java
@@ -1,0 +1,64 @@
+package com.bakdata.conquery.sql.conquery;
+
+import java.util.List;
+import java.util.UUID;
+import java.util.stream.Stream;
+
+import com.bakdata.conquery.apiv1.query.QueryDescription;
+import com.bakdata.conquery.models.auth.entities.User;
+import com.bakdata.conquery.models.config.ConqueryConfig;
+import com.bakdata.conquery.models.datasets.Dataset;
+import com.bakdata.conquery.models.execution.InternalExecution;
+import com.bakdata.conquery.models.execution.ManagedExecution;
+import com.bakdata.conquery.models.query.ExecutionManager;
+import com.bakdata.conquery.models.query.results.EntityResult;
+import com.bakdata.conquery.models.query.results.ShardResult;
+import com.bakdata.conquery.models.worker.Namespace;
+import com.google.common.cache.RemovalNotification;
+
+public class SqlExecutionManager implements ExecutionManager {
+	@Override
+	public ManagedExecution runQuery(Namespace namespace, QueryDescription query, User user, Dataset submittedDataset, ConqueryConfig config, boolean system) {
+		return null;
+	}
+
+	@Override
+	public void execute(Namespace namespace, ManagedExecution execution, ConqueryConfig config) {
+
+	}
+
+	@Override
+	public ManagedExecution createExecution(QueryDescription query, User user, Dataset submittedDataset, boolean system) {
+		return null;
+	}
+
+	@Override
+	public ManagedExecution createQuery(QueryDescription query, UUID queryId, User user, Dataset submittedDataset, boolean system) {
+		return null;
+	}
+
+	@Override
+	public void cancelQuery(Dataset dataset, ManagedExecution query) {
+
+	}
+
+	@Override
+	public <R extends ShardResult, E extends ManagedExecution & InternalExecution<R>> void handleQueryResult(R result) {
+
+	}
+
+	@Override
+	public void addQueryResult(ManagedExecution execution, List<EntityResult> queryResults) {
+
+	}
+
+	@Override
+	public void clearQueryResults(ManagedExecution execution) {
+
+	}
+
+	@Override
+	public Stream<EntityResult> streamQueryResults(ManagedExecution execution) {
+		return null;
+	}
+}
