+++ backend/src/main/java/com/bakdata/conquery/sql/conquery/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Integration with the existing ConQuery architecture.
+ */
+package com.bakdata.conquery.sql.conquery;
