+++ backend/src/test/java/com/bakdata/conquery/integration/json/JsonIntegrationTest.java
@@ -48,8 +48,7 @@ public class JsonIntegrationTest extends IntegrationTest.Simple {
 
 		//ensure the metadata is collected
 
-
-		conquery.getNamespace().sendToAll(new UpdateMatchingStatsMessage(conquery.getNamespace().getStorage().getAllConcepts()));
+		conquery.getNamespace().getWorkerHandler().sendToAll(new UpdateMatchingStatsMessage(conquery.getNamespace().getStorage().getAllConcepts()));
 
 		conquery.waitUntilWorkDone();
